+++ MeasurementsApi/Migrations/DataContextModelSnapshot.cs
@@ -45,8 +45,8 @@ namespace MeasurementsApi.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<double>("PricePerUnit")
-                        .HasColumnType("float");
+                    b.Property<int>("PricePerUnit")
+                        .HasColumnType("int");
 
                     b.Property<int>("Quantity")
                         .HasColumnType("int");
